Add files via upload
The YouTube Clone in ReactJS is a simplified version of the YouTube platform, allowing users to search for videos, watch them, and see recommendations. This clone utilizes the YouTube API to fetch data for videos, channels, and search queries. It also includes essential features of YouTube, such as browsing categories, viewing details of each video, and loading related videos.
Tech :
1. ReactJS for building the UI components.
2. React Router for handling navigation between pages.
3.YouTube Data API v3 for fetching video, channel, and search data.
4.Styled Components for styling the components and making the app visually appealing.
5. Fetch API for making API requests.